requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
py-tgcalls==0.9.7
pytube==15.0.0
yt-dlp==2023.12.30
pymongo==4.6.1
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
youtube-search-python==1.6.6
spotipy==2.23.0
speech_recognition==3.10.1
pydub==0.25.1
ffmpeg-python==0.2.0
psutil==5.9.6
motor==3.3.2
dnspython==2.4.2
